+++ src/main/java/roomescape/common/exception/CustomException.java
@@ -1,6 +1,7 @@
 package roomescape.common.exception;
 
 public abstract class CustomException extends RuntimeException {
+
     public CustomException(String message) {
         super(message);
     }
