+++ src/main/java/roomescape/common/exception/DuplicatedException.java
@@ -1,6 +1,7 @@
 package roomescape.common.exception;
 
 public class DuplicatedException extends CustomException {
+
     public DuplicatedException(String message) {
         super(message);
     }
