+++ src/main/java/roomescape/common/exception/NotFoundException.java
@@ -1,6 +1,7 @@
 package roomescape.common.exception;
 
 public class NotFoundException extends CustomException {
+
     public NotFoundException(String message) {
         super(message);
     }
