+++ src/main/java/roomescape/common/exception/OperationNotAllowedException.java
@@ -1,6 +1,7 @@
 package roomescape.common.exception;
 
 public class OperationNotAllowedException extends CustomException {
+
     public OperationNotAllowedException(String message) {
         super(message);
     }
