+++ src/main/java/roomescape/common/exception/PaymentClientException.java
@@ -1,6 +1,7 @@
 package roomescape.common.exception;
 
 public class PaymentClientException extends CustomException {
+
     public PaymentClientException(String message) {
         super(message);
     }
