+++ src/main/java/roomescape/common/exception/PaymentServerException.java
@@ -1,6 +1,7 @@
 package roomescape.common.exception;
 
 public class PaymentServerException extends CustomException {
+
     public PaymentServerException(String message) {
         super(message);
     }
