+++ src/main/java/roomescape/common/exception/ResourceInUseException.java
@@ -1,6 +1,7 @@
 package roomescape.common.exception;
 
 public class ResourceInUseException extends CustomException {
+
     public ResourceInUseException(String message) {
         super(message);
     }
