+++ src/main/java/roomescape/common/exception/UnauthorizedException.java
@@ -1,6 +1,7 @@
 package roomescape.common.exception;
 
 public class UnauthorizedException extends CustomException {
+
     public UnauthorizedException(String message) {
         super(message);
     }
